fix(demultiplexer): fix buildifier, lint_components, and missing FakeSamplerMock in def
- Fix buildifier formatting in impl/BUILD.bazel (sort srcs and deps alphabetically)
- Add comp/aggregator/demultiplexer to ignore_fx_import in tasks/components.py
  (demultiplexer_fake_sampler_mock.go has //go:build test but lint treats it as impl)
- Add comp/aggregator/demultiplexer to ignore_provide_component_constructor_missing
  (fx.go delegates to impl.Module which uses an unexported constructor)
- Add def/component_mock.go with Mock and FakeSamplerMock interfaces (moved from
  root component_mock.go); impl/demultiplexer_fake_sampler_mock.go references
  demultiplexerComp.FakeSamplerMock from def/ which was missing
- Update def/BUILD.bazel to include component_mock.go and pkg/metrics dep
- Simplify mock/mock.go and mock/BUILD.bazel (remove duplicate type declarations
  now that they live in def/)
