+++ comp/aggregator/demultiplexer/def/BUILD.bazel
@@ -2,12 +2,16 @@ load("@rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "def",
-    srcs = ["component.go"],
+    srcs = [
+        "component.go",
+        "component_mock.go",
+    ],
     importpath = "github.com/DataDog/datadog-agent/comp/aggregator/demultiplexer/def",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/aggregator",
         "//pkg/aggregator/sender",
+        "//pkg/metrics",
         "//pkg/serializer",
     ],
 )
