+++ comp/aggregator/demultiplexer/impl/BUILD.bazel
@@ -39,11 +39,11 @@ go_library(
 go_test(
     name = "impl_test",
     srcs = [
-        "demultiplexer_mock_test.go",
-        "demultiplexer_mock.go",
         "demultiplexer_fake_sampler_mock.go",
-        "test_agent_demultiplexer.go",
+        "demultiplexer_mock.go",
+        "demultiplexer_mock_test.go",
         "status_test.go",
+        "test_agent_demultiplexer.go",
     ],
     embed = [":impl"],
     gotags = ["test"],
@@ -58,6 +58,7 @@ go_test(
         "//comp/core/status",
         "//comp/core/tagger/def",
         "//comp/core/tagger/fx-mock",
+        "//comp/core/tagger/impl-noop",
         "//comp/filterlist/fx-mock",
         "//comp/filterlist/impl",
         "//comp/forwarder/defaultforwarder",
@@ -69,7 +70,6 @@ go_test(
         "//comp/serializer/logscompression/fx-mock",
         "//comp/serializer/metricscompression/def",
         "//comp/serializer/metricscompression/fx-mock",
-        "//comp/core/tagger/impl-noop",
         "//pkg/aggregator",
         "//pkg/aggregator/mocksender",
         "//pkg/aggregator/sender",
