+++ tasks/components.py
@@ -179,6 +179,7 @@ components_missing_implementation_folder = [
 ]
 
 ignore_fx_import = [
+    "comp/aggregator/demultiplexer",
     "comp/otelcol/logsagentpipeline",
     "comp/core/workloadmeta",
     "comp/rdnsquerier",
@@ -187,6 +188,7 @@ ignore_fx_import = [
 ]
 
 ignore_provide_component_constructor_missing = [
+    "comp/aggregator/demultiplexer",
     "comp/otelcol/logsagentpipeline",
     "comp/core/workloadmeta",
     "comp/trace/agent",
